Ai-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
bleak==0.22.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
deprecation==2.1.0
Django==5.1.7
django-cors-headers==4.7.0
django-restframework==0.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.5.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
gotrue==2.12.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
multidict==6.2.0
oauthlib==3.2.2
packaging==24.2
pillow==11.2.1
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.9.0
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
winrt-runtime==2.3.0
winrt-Windows.Devices.Bluetooth==2.3.0
winrt-Windows.Devices.Bluetooth.Advertisement==2.3.0
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==2.3.0
winrt-Windows.Devices.Enumeration==2.3.0
winrt-Windows.Foundation==2.3.0
winrt-Windows.Foundation.Collections==2.3.0
winrt-Windows.Storage.Streams==2.3.0
yarl==1.18.3
